Move API traits to separate package
